Add --one-path-per-blob
This causes the blob_shas_and_paths.txt to be created with one pathname
per line, instead of grouping all the paths for a specific blob on a
single line.

This supports the pathbase feature in the next commit.

This means that occasionally a blob's sha will be repeated in the file.
This is expected and normal.

Signed-off-by: Stuart MacDonald <stuartm.coding@gmail.com>